backend/requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
Werkzeug==3.0.1
